+++ excluirTabela.c
@@ -1,0 +1,44 @@
+#include "buffend.h"
+
+/* ---------------------------------------------------------------------------------------------- 
+    Objetivo:   Função para exclusão de tabelas.
+    Parametros: Nome da tabela (char).    
+    Retorno:    INT
+                SUCCESS, 
+                ERRO_REMOVER_ARQUIVO_OBJECT, 
+                ERRO_REMOVER_ARQUIVO_SCHEMA, 
+                ERRO_LEITURA_DADOS.
+   ---------------------------------------------------------------------------------------------*/
+
+int excluirTabela(char *nomeTabela){
+    struct fs_objects objeto;
+    tp_table *esquema;
+    int x,erro;
+    char str[20]; 
+    char dat[5] = ".dat";
+
+    strcpy (str, nomeTabela); 
+    strcat (str, dat);              //Concatena e junta o nome com .dat
+
+    abreTabela(nomeTabela, &objeto, &esquema);
+    tp_buffer *bufferpoll = initbuffer();
+
+    if(bufferpoll == ERRO_DE_ALOCACAO){
+        printf("Erro ao alocar memória para o buffer.\n");
+        return ERRO_LEITURA_DADOS;
+    }
+
+    erro = SUCCESS;
+    for(x = 0; erro == SUCCESS; x++)
+        erro = colocaTuplaBuffer(bufferpoll, x, esquema, objeto);        
+    
+    if(procuraSchemaArquivo(objeto) != 0)
+        return ERRO_REMOVER_ARQUIVO_SCHEMA;
+
+    if(procuraObjectArquivo(nomeTabela) != 0)
+       return ERRO_REMOVER_ARQUIVO_OBJECT;
+        
+    remove(str);
+    
+    return SUCCESS;
+}
