+++ existeArquivo.c
@@ -1,0 +1,18 @@
+#include "buffend.h"
+
+/* ---------------------------------------------------------------------------------------------- 
+    Objetivo:   Verificação de existência de um arquivo.
+    Parametros: Nome do arquivo.    
+    Retorno:    INT 1 (existe) , 0 (não existe).
+   ---------------------------------------------------------------------------------------------*/
+    
+int existeArquivo(const char* filename){
+    FILE* fptr = fopen(filename, "r");
+    if (fptr != NULL){
+        fclose(fptr);
+        
+        return 1;
+    }
+
+    return 0;
+}
