+++ getTupla.c
@@ -1,0 +1,28 @@
+#include "buffend.h"
+
+// INSERE UMA TUPLA NO BUFFER!
+char *getTupla(tp_table *campos,struct fs_objects objeto, int from) { //Pega uma tupla do disco a partir do valor de from
+
+    int tamTpl = tamTupla(campos, objeto);
+    char *linha=(char *)malloc(sizeof(char)*tamTpl);
+    FILE *dados;
+
+    from = from * tamTpl;
+
+    dados = fopen(objeto.nArquivo, "r");
+
+    if (dados == NULL)
+        return ERRO_DE_LEITURA;
+
+    fseek(dados, from, 1);
+    if(fgetc (dados) != EOF){
+        fseek(dados, -1, 1);
+        fread(linha, sizeof(char), tamTpl, dados); //Traz a tupla inteira do arquivo
+    }
+    else{       //Caso em que o from possui uma valor inválido para o arquivo de dados
+        fclose(dados);
+        return ERRO_DE_LEITURA;
+    }
+    fclose(dados);
+    return linha;
+}
