+++ procuraAtributoFK.c
@@ -1,0 +1,52 @@
+#include "buffend.h"
+
+/* ---------------------------------------------------------------------------------------------- 
+    Objetivo:   Retorna vetor de esquemas com todos os atributos chaves (PK, FK e NPK)
+    Parametros: Objeto da tabela.
+    Retorno:    Vetor de esquemas vetEsqm
+   ---------------------------------------------------------------------------------------------*/
+
+tp_table *procuraAtributoFK(struct fs_objects objeto){
+    FILE *schema;
+    int cod = 0, chave, i = 0;
+    char *tupla = (char *)malloc(sizeof(char) * 109);
+    tp_table *esquema = (tp_table *)malloc(sizeof(tp_table)*objeto.qtdCampos);
+    tp_table *vetEsqm = (tp_table *)malloc(sizeof(tp_table)*objeto.qtdCampos);
+
+    if((schema = fopen("fs_schema.dat", "a+b")) == NULL){
+        printf("Erro GRAVE ao abrir o ESQUEMA.\nAbortando...\n");
+        exit(1);
+    }
+    
+    while((fgetc (schema) != EOF) && i < objeto.qtdCampos){ // Varre o arquivo ate encontrar todos os campos com o codigo da tabela.
+        fseek(schema, -1, 1);
+
+        if(fread(&cod, sizeof(int), 1, schema)){ // Le o codigo da tabela.
+            if(cod == objeto.cod){
+                fread(tupla, sizeof(char), TAMANHO_NOME_CAMPO, schema);
+                strcpy(esquema[i].nome,tupla);                  // Copia dados do campo para o esquema.
+
+                fread(&esquema[i].tipo , sizeof(char), 1 , schema);      
+                fread(&esquema[i].tam  , sizeof(int) , 1 , schema);   
+                fread(&chave, sizeof(int) , 1 , schema);  
+
+                fread(tupla, sizeof(char), TAMANHO_NOME_TABELA, schema);
+                strcpy(esquema[i].tabelaApt,tupla);
+
+                fread(tupla, sizeof(char), TAMANHO_NOME_CAMPO, schema);           
+                strcpy(esquema[i].attApt,tupla);
+                
+                strcpy(vetEsqm[i].tabelaApt, esquema[i].tabelaApt);
+                strcpy(vetEsqm[i].attApt, esquema[i].attApt);
+                strcpy(vetEsqm[i].nome, esquema[i].nome);
+                vetEsqm[i].chave = chave;
+
+                i++;
+            } else {
+                fseek(schema, 109, 1);
+            }
+        }
+    }
+
+    return vetEsqm;
+}
